pyproject.toml
[project]
name = "delta-optim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sebastien VEY", email = "sebvey@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "delta-spark>=3.2.1",
    "deltalake>=0.22.0",
    "duckdb>=1.1.3",
    "marimo>=0.9.27",
    "numpy>=2.1.3",
    "polars>=1.15.0",
    "pyspark>=3.5.3",
    "timeit-decorator>=1.1.1",
    "xfp>=1.0.0",
]

[project.scripts]
create-delta = "delta_optim.job_create_delta:main"
query-delta = "delta_optim.job_query_delta:main"
compact-delta = "delta_optim.job_compact_delta:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
